added procces cpu util

+++ src/process.cpp
@@ -26,7 +26,7 @@ float Process::CpuUtilization(){
     unsigned long long starttime=std::stoull(vs[4]);
     std::cerr<<utime<<" "<<stime<<" "<<cutime<<" "<<cstime<<" "<<starttime<<" "<<LinuxParser::UpTime()<<std::endl;
     unsigned long long total_time = utime + stime + cutime + cstime;
-    long seconds = LinuxParser::UpTime() - (starttime / Hertz);
+    long seconds = LinuxParser::UpTime() - (float)(starttime / Hertz);
     return (float)((total_time / Hertz) / seconds);
 }
